backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==3.14.0